rearchitect wifictl
This change moves wifictl to a one-file-per-network architecture.
Doing so means it has a dramatically simpler deletion routine which
is already a large benefit; however, the big win is that it enables
a feature it's never had before: the ability to connect to a
specified network.

This is a really simple feature, but because wifictl rather
intentionally avoids touching wpa_cli, there's actually no simple
way to do this with a single config file. This commit offers a simple
solution: have a config file per-network, and materialize the unified
config file when-instructed. Connecting to a specific network is as
easy as adding a very-high priority marker to that network when
materializing the unified config, then reloading.

This change also technically enables a lot of other features that
aren't currently implemented like temporarily disabling networks.

Signed-off-by: Sam Stuewe <samuel.stuewe@gmail.com>